# Use sync.Pool for gzip writer