By using a pool of gzip.Writers, at the given compression level, we are able to realize significant improvements to memory usage, allocations (and thus GC) and CPU time. When a writer is retrieved from the pool, the `Reset` method is called to re-use the underlying buffers, while pointing them at the new ResponseWriter.

I have created a benchmark by which I gathered data from before and after my changes:

```
go test -v -run=^$ -bench=. -benchtime=10s -memprofile=prof.mem


benchmark               old ns/op     new ns/op     delta
Benchmark_ServeHTTP     187922        95292         -49.29%

benchmark               old allocs     new allocs     delta
Benchmark_ServeHTTP     62             38             -38.71%

benchmark               old bytes     new bytes     delta
Benchmark_ServeHTTP     1457960       4145          -99.72%
```

In addition, when benchmarking this on my project, my memory usage dropped from 3.2GB to 511MB.

Please let me know if there are any changes that need to be made before it is considered ready to merge.

Thank you!